+++ src/lib/row/GroupRow.js
@@ -6,25 +6,37 @@ class GroupRow extends Component {
   static propTypes = {
     onClick: PropTypes.func.isRequired,
     onDoubleClick: PropTypes.func.isRequired,
+    onContextMenu: PropTypes.func.isRequired,
     isEvenRow: PropTypes.bool.isRequired,
     style: PropTypes.object.isRequired,
-    clickTolerance: PropTypes.number.isRequired
+    clickTolerance: PropTypes.number.isRequired,
+    group: PropTypes.object.isRequired,
+    horizontalLineClassNamesForGroup: PropTypes.func
   }
 
   render() {
     const {
+      onContextMenu,
       onDoubleClick,
       isEvenRow,
       style,
       onClick,
-      clickTolerance
+      clickTolerance,
+      horizontalLineClassNamesForGroup,
+      group
     } = this.props
+
+    let classNamesForGroup = [];
+    if (horizontalLineClassNamesForGroup) {
+      classNamesForGroup = horizontalLineClassNamesForGroup(group);
+    }
 
     return (
       <PreventClickOnDrag clickTolerance={clickTolerance} onClick={onClick}>
         <div
+          onContextMenu={onContextMenu}
           onDoubleClick={onDoubleClick}
-          className={isEvenRow ? 'rct-hl-even' : 'rct-hl-odd'}
+          className={(isEvenRow ? 'rct-hl-even ' : 'rct-hl-odd ') + (classNamesForGroup ? classNamesForGroup.join(' ') : '')}
           style={style}
         />
       </PreventClickOnDrag>